refactor: 타입 정의, 사용하는 곳에 가깝게 두기
