+++ src/App.tsx
@@ -2,7 +2,7 @@ import "./index.css";
 import "./reset.css";
 import Header from "./components/header/Header";
 import ProductContainer from "./components/productContainer/ProductContainer";
-import useFetchCartProducts from "./hooks/useFetchCartProducts";
+import useFetchCartProducts from "./hooks/useFetchCartProducts/useFetchCartProducts";
 
 function App() {
   const { cartItemIds, setCartItemIds, fetchCartProducts } =
