+++ src/components/cartToggleButton/useCartToggleButton.ts
@@ -1,4 +1,4 @@
-import { useToast } from "../../hooks/useToast";
+import { useToast } from "../../hooks/useToast/useToast";
 import request from "../../utils/request";
 
 const MAX_CART_AMOUNT = 50;
