+++ src/components/productCardList/ProductCardList.tsx
@@ -1,7 +1,7 @@
 import ProductCard from "../productCard/ProductCard";
 import { CardListContainer } from "./ProductCardList.css";
 import { ProductPageResponse } from "../../types/response.types";
-import useFetchProducts from "../../hooks/useFetchProducts";
+import useFetchProducts from "../../hooks/useFetchProducts/useFetchProducts";
 import { categoryType, sortType } from "../../types/index.types";
 import ProductCardListSkeleton from "../productCardListSkeleton/ProductCardListSkeleton";
