+++ src/hooks/useToast/useToast.ts
@@ -1,5 +1,5 @@
 import { useContext } from "react";
-import { ToastContext } from "../context/toastContext";
+import { ToastContext } from "../../context/toastContext";
 
 export const useToast = () => {
   const context = useContext(ToastContext);
