+++ src/types/response.types.ts
@@ -1,4 +1,3 @@
-// 상품 개별 항목 타입
 export interface Product {
   id: number;
   name: string;
@@ -6,21 +5,12 @@ export interface Product {
   imageUrl: string;
   category: string;
 }
-
-export interface CartItem {
-  id: number;
-  product: Product;
-  quantity: number;
-}
-
-// 정렬 정보 타입
 export interface Sort {
   empty: boolean;
   sorted: boolean;
   unsorted: boolean;
 }
 
-// 페이지네이션 정보 타입
 export interface Pageable {
   offset: number;
   pageNumber: number;
@@ -28,39 +18,4 @@ export interface Pageable {
   paged: boolean;
   unpaged: boolean;
   sort: Sort;
-}
-
-// 전체 응답 타입
-export interface ProductPageResponse {
-  content: Product[];
-  pageable: Pageable;
-  totalElements: number;
-  totalPages: number;
-  last: boolean;
-  first: boolean;
-  size: number;
-  number: number;
-  numberOfElements: number;
-  empty: boolean;
-  sort: Sort;
-}
-
-export interface CartItem {
-  id: number;
-  quantity: number;
-  product: Product;
-}
-
-export interface CartProduct {
-  content: CartItem[];
-  empty: boolean;
-  first: boolean;
-  last: boolean;
-  number: number;
-  numberOfElements: number;
-  pageable: Pageable;
-  size: number;
-  sort: Sort;
-  totalElements: number;
-  totalPages: number;
 }
